Refactored test guard, added bias guard in hipblasgemm
